VB-3739 created by field for blocking date (#823)
